Add CitySelector
Took 37 minutes
