+++ DentalCore.Wpf/Views/Components/CitySelector.xaml.cs
@@ -1,0 +1,11 @@
+﻿using System.Windows.Controls;
+
+namespace DentalCore.Wpf.Views.Components;
+
+public partial class CitySelector : UserControl
+{
+    public CitySelector()
+    {
+        InitializeComponent();
+    }
+}
